Initial commit
Pretty dumb but does a decent job at gathering files
and organizing them into a group of maven components
that Sonatype Nexus will happily consume. The premise is
that snapshots have multiple versions and artifacts for
each "snapshot", while there is only a few artifacts for
a specific release version.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>